refactor(iocuak): extract container module metadata logic to containerModuleMetadata folder
