+++ packages/iocuak/src/container/modules/api/ContainerApi.int.spec.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata';
 
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
-import { ContainerModuleMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleMetadataApi';
+import { ContainerModuleMetadataApi } from '../../../containerModuleMetadata/models/api/ContainerModuleMetadataApi';
 import { inject } from '../../../metadata/decorators/inject';
 import { injectable } from '../../../metadata/decorators/injectable';
 import { ContainerModuleBindingServiceApi } from '../../services/api/ContainerModuleBindingServiceApi';
