+++ packages/iocuak/src/container/services/api/ContainerModuleServiceApi.ts
@@ -1,5 +1,5 @@
 import { ContainerModuleApi } from '../../../containerModule/models/api/ContainerModuleApi';
-import { ContainerModuleMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleMetadataApi';
+import { ContainerModuleMetadataApi } from '../../../containerModuleMetadata/models/api/ContainerModuleMetadataApi';
 
 export interface ContainerModuleServiceApi {
   load(containerModuleApi: ContainerModuleApi): void;
