+++ packages/iocuak/src/container/services/api/ContainerServiceApiImplementation.spec.ts
@@ -1,5 +1,5 @@
 jest.mock(
-  '../../../containerModuleTask/utils/api/convertToContainerModuleMetadata',
+  '../../../containerModuleMetadata/utils/api/convertToContainerModuleMetadata',
 );
 jest.mock('../../utils/bind');
 jest.mock('../../utils/bindToValue');
@@ -8,9 +8,9 @@ jest.mock('../../../metadata/utils/api/convertBindingToBindingApi');
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
 import { ContainerModuleApi } from '../../../containerModule/models/api/ContainerModuleApi';
-import { ContainerModuleMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleMetadataApi';
-import { ContainerModuleMetadata } from '../../../containerModuleTask/models/domain/ContainerModuleMetadata';
-import { convertToContainerModuleMetadata } from '../../../containerModuleTask/utils/api/convertToContainerModuleMetadata';
+import { ContainerModuleMetadataApi } from '../../../containerModuleMetadata/models/api/ContainerModuleMetadataApi';
+import { ContainerModuleMetadata } from '../../../containerModuleMetadata/models/domain/ContainerModuleMetadata';
+import { convertToContainerModuleMetadata } from '../../../containerModuleMetadata/utils/api/convertToContainerModuleMetadata';
 import { BindingApi } from '../../../metadata/models/api/BindingApi';
 import { BindingTypeApi } from '../../../metadata/models/api/BindingTypeApi';
 import { Binding } from '../../../metadata/models/domain/Binding';
