+++ packages/iocuak/src/container/services/cuaktask/ContainerModuleServiceImplementation.ts
@@ -2,9 +2,9 @@ import * as cuaktask from '@cuaklabs/cuaktask';
 
 import { Builder } from '../../../common/modules/domain/Builder';
 import { ContainerModule } from '../../../containerModule/models/domain/ContainerModule';
+import { ContainerModuleMetadata } from '../../../containerModuleMetadata/models/domain/ContainerModuleMetadata';
 import { ContainerModuleLoadFromMetadataTask } from '../../../containerModuleTask/models/cuaktast/ContainerModuleLoadFromMetadataTask';
 import { ContainerModuleLoadFromMetadataTaskKind } from '../../../containerModuleTask/models/domain/ContainerModuleLoadFromMetadataTaskKind';
-import { ContainerModuleMetadata } from '../../../containerModuleTask/models/domain/ContainerModuleMetadata';
 import { ContainerModuleTaskKind } from '../../../containerModuleTask/models/domain/ContainerModuleTaskKind';
 import { ContainerModuleTaskKindType } from '../../../containerModuleTask/models/domain/ContainerModuleTaskKindType';
 import { ContainerModuleService } from '../domain/ContainerModuleService';
