+++ packages/iocuak/src/containerModuleMetadata/models/domain/ContainerModuleClassMetadata.ts
