+++ packages/iocuak/src/containerModuleMetadata/models/domain/ContainerModuleMetadata.ts
