+++ packages/iocuak/src/containerModuleMetadata/models/domain/ContainerModuleMetadataBase.ts
