+++ packages/iocuak/src/containerModuleMetadata/models/domain/ContainerModuleMetadataType.ts
