+++ packages/iocuak/src/e2e/definitions/container/containerDefinitions.ts
@@ -6,8 +6,8 @@ import sinonChai from 'sinon-chai';
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
 import { ContainerApi } from '../../../container/modules/api/ContainerApi';
-import { ContainerModuleClassMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleClassMetadataApi';
-import { ContainerModuleFactoryMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleFactoryMetadataApi';
+import { ContainerModuleClassMetadataApi } from '../../../containerModuleMetadata/models/api/ContainerModuleClassMetadataApi';
+import { ContainerModuleFactoryMetadataApi } from '../../../containerModuleMetadata/models/api/ContainerModuleFactoryMetadataApi';
 import { BindingApi } from '../../../metadata/models/api/BindingApi';
 import { BindingTypeApi } from '../../../metadata/models/api/BindingTypeApi';
 import { ClassMetadataApi } from '../../../metadata/models/api/ClassMetadataApi';
