+++ packages/iocuak/src/e2e/definitions/container/parameters/containerModuleMetadata/getContainerModuleMetadataWithContainerModuleClassMetadataImports.ts
@@ -3,7 +3,7 @@ import sinon from 'sinon';
 import { Newable } from '../../../../../common/models/domain/Newable';
 import { ContainerModuleBindingServiceApi } from '../../../../../container/services/api/ContainerModuleBindingServiceApi';
 import { ContainerModuleApi } from '../../../../../containerModule/models/api/ContainerModuleApi';
-import { ContainerModuleClassMetadataApi } from '../../../../../containerModuleTask/models/api/ContainerModuleClassMetadataApi';
+import { ContainerModuleClassMetadataApi } from '../../../../../containerModuleMetadata/models/api/ContainerModuleClassMetadataApi';
 import { inject } from '../../../../../metadata/decorators/inject';
 import { injectable } from '../../../../../metadata/decorators/injectable';
 import { getTypeServiceWithNoDependenciesParameter } from '../../../common/parameters/typeService/getTypeServiceWithNoDependenciesParameter';
